refactor the regitry code
as far as I understood it
