+++ src/main/kotlin/org/ender_development/catalyx/core/ICatalyxMod.kt
@@ -2,10 +2,10 @@ package org.ender_development.catalyx.core
 
 import net.minecraft.creativetab.CreativeTabs
 import net.minecraftforge.fml.common.Mod
+import org.ender_development.catalyx.api.v1.registry.IBlockProvider
+import org.ender_development.catalyx.api.v1.registry.IItemProvider
 import org.ender_development.catalyx.core.registry.CatalyxBlockRegistry
 import org.ender_development.catalyx.core.registry.CatalyxItemRegistry
-import org.ender_development.catalyx.core.registry.IBlockProvider
-import org.ender_development.catalyx.core.registry.IItemProvider
 
 interface ICatalyxMod {
 	 companion object {
