+++ src/main/kotlin/org/ender_development/catalyx/core/blocks/BaseBlock.kt
@@ -14,9 +14,9 @@ import net.minecraft.util.math.BlockPos
 import net.minecraft.world.IBlockAccess
 import net.minecraftforge.client.model.ModelLoader
 import net.minecraftforge.event.RegistryEvent
+import org.ender_development.catalyx.api.v1.registry.IBlockProvider
 import org.ender_development.catalyx.core.ICatalyxMod
 import org.ender_development.catalyx.core.register
-import org.ender_development.catalyx.core.registry.IBlockProvider
 import org.ender_development.catalyx.core.utils.SideUtils
 
 /**
