+++ src/main/kotlin/org/ender_development/catalyx/core/registry/CatalyxBlockRegistry.kt
@@ -6,13 +6,15 @@ import net.minecraftforge.event.RegistryEvent
 import net.minecraftforge.fml.common.Mod
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent
 import org.ender_development.catalyx.Catalyx
+import org.ender_development.catalyx.api.v1.registry.IBlockProvider
+import org.ender_development.catalyx.api.v1.registry.ICatalyxRegistry
 import org.ender_development.catalyx.core.Reference
 import org.ender_development.catalyx.core.utils.DevUtils
 
 @Suppress("unused")
 @Mod.EventBusSubscriber(modid = Reference.MODID)
 object CatalyxBlockRegistry : ICatalyxRegistry<Block, IBlockProvider> {
-	override val registry = CatalyxRegistry<IBlockProvider>()
+	override val registry = CatalyxProviderRegistry<IBlockProvider>()
 
 	@SubscribeEvent
 	override fun register(event: RegistryEvent.Register<Block>) {
