+++ src/main/kotlin/org/ender_development/catalyx/core/registry/CatalyxProviderRegistry.kt
@@ -1,23 +1,30 @@
 package org.ender_development.catalyx.core.registry
 
 import net.minecraft.util.ResourceLocation
+import org.ender_development.catalyx.api.v1.registry.ICatalyxProviderRegistry
+import org.ender_development.catalyx.api.v1.registry.IProvider
 import org.ender_development.catalyx.core.utils.extensions.modLoaded
 
-class CatalyxRegistry<V : IProvider<*>> : HashMap<ResourceLocation, Pair<V, Boolean>>() {
-	fun add(provider: V): Boolean =
+/**
+ * Collection of all [IProvider] of a given type
+ *
+ * @param V the Type of the Provider for the things that should be managed here
+ */
+class CatalyxProviderRegistry<V : IProvider<*>> :
+	HashMap<ResourceLocation, Pair<V, Boolean>>(),
+	ICatalyxProviderRegistry<V> {
+
+	override fun add(provider: V): Boolean =
 		provider.instance.registryName?.let {
 			this[it] = provider to (provider.isEnabled() && provider.modDependencies.evaluateModDependencies())
 			true
 		} ?: false
 
-	val enabled: List<V>
-		get() = values.filter { it.second }.map { it.first }
-
-	val plural
-		get() = if(enabled.size == 1) "" else "s"
-
 	/**
 	 * Special expansion to evaluate modDependencies strings.
+	 *
+	 * @return true if valid
+	 * @see [IProvider.modDependencies]
 	 */
 	fun String.evaluateModDependencies(): Boolean =
 		isEmpty() || split(',', ';').all {
